refactoring and fixing Unit tests

+++ src/main/java/com/sogeti/java/anonymous_artist/response/ProductResponse.java
@@ -12,7 +12,7 @@ public record ProductResponse(
         String smallSummary,
         String description,
         BigDecimal price,
-        Long amountInStock
+        Integer amountInStock
 ) {
 
 }
